[To rel/1.0] Rename StorageEngineV2 to StorageEngine (#8238)
