+++ server/src/main/java/org/apache/iotdb/db/consensus/DataRegionConsensusImpl.java
@@ -31,7 +31,7 @@ import org.apache.iotdb.consensus.config.RatisConfig.Snapshot;
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.consensus.statemachine.DataRegionStateMachine;
-import org.apache.iotdb.db.engine.StorageEngineV2;
+import org.apache.iotdb.db.engine.StorageEngine;
 
 import org.apache.ratis.util.SizeInBytes;
 import org.apache.ratis.util.TimeDuration;
@@ -164,7 +164,7 @@ public class DataRegionConsensusImpl {
                       .build(),
                   gid ->
                       new DataRegionStateMachine(
-                          StorageEngineV2.getInstance().getDataRegion((DataRegionId) gid)))
+                          StorageEngine.getInstance().getDataRegion((DataRegionId) gid)))
               .orElseThrow(
                   () ->
                       new IllegalArgumentException(
