+++ server/src/main/java/org/apache/iotdb/db/query/executor/AggregationExecutor.java
@@ -24,12 +24,9 @@ import org.apache.iotdb.commons.path.AlignedPath;
 import org.apache.iotdb.commons.path.MeasurementPath;
 import org.apache.iotdb.commons.path.PartialPath;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
-import org.apache.iotdb.db.engine.StorageEngine;
 import org.apache.iotdb.db.engine.querycontext.QueryDataSource;
-import org.apache.iotdb.db.engine.storagegroup.DataRegion;
 import org.apache.iotdb.db.exception.StorageEngineException;
 import org.apache.iotdb.db.exception.query.QueryProcessException;
-import org.apache.iotdb.db.metadata.utils.MetaUtils;
 import org.apache.iotdb.db.qp.physical.crud.AggregationPlan;
 import org.apache.iotdb.db.qp.physical.crud.QueryPlan;
 import org.apache.iotdb.db.qp.physical.crud.RawDataQueryPlan;
@@ -53,20 +50,15 @@ import org.apache.iotdb.tsfile.read.common.BatchData;
 import org.apache.iotdb.tsfile.read.common.IBatchDataIterator;
 import org.apache.iotdb.tsfile.read.common.RowRecord;
 import org.apache.iotdb.tsfile.read.expression.IExpression;
-import org.apache.iotdb.tsfile.read.expression.impl.GlobalTimeExpression;
 import org.apache.iotdb.tsfile.read.filter.basic.Filter;
 import org.apache.iotdb.tsfile.read.query.dataset.QueryDataSet;
 import org.apache.iotdb.tsfile.read.query.timegenerator.TimeGenerator;
-import org.apache.iotdb.tsfile.utils.Pair;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -107,61 +99,7 @@ public class AggregationExecutor {
   /** execute aggregate function with only time filter or no filter. */
   public QueryDataSet executeWithoutValueFilter(AggregationPlan aggregationPlan)
       throws StorageEngineException, IOException, QueryProcessException {
-
-    Filter timeFilter = null;
-    if (expression != null) {
-      timeFilter = ((GlobalTimeExpression) expression).getFilter();
-    }
-
-    // TODO use multi-thread
-    Map<PartialPath, List<Integer>> pathToAggrIndexesMap =
-        MetaUtils.groupAggregationsBySeries(selectedSeries);
-    // Attention: this method will REMOVE aligned path from pathToAggrIndexesMap
-    Map<AlignedPath, List<List<Integer>>> alignedPathToAggrIndexesMap =
-        MetaUtils.groupAlignedSeriesWithAggregations(pathToAggrIndexesMap);
-
-    List<PartialPath> groupedPathList =
-        new ArrayList<>(pathToAggrIndexesMap.size() + alignedPathToAggrIndexesMap.size());
-    groupedPathList.addAll(pathToAggrIndexesMap.keySet());
-    groupedPathList.addAll(alignedPathToAggrIndexesMap.keySet());
-
-    // TODO-Cluster: group the paths by database to reduce communications
-    Pair<List<DataRegion>, Map<DataRegion, List<PartialPath>>> lockListAndProcessorToSeriesMapPair =
-        StorageEngine.getInstance().mergeLock(groupedPathList);
-    List<DataRegion> lockList = lockListAndProcessorToSeriesMapPair.left;
-    Map<DataRegion, List<PartialPath>> processorToSeriesMap =
-        lockListAndProcessorToSeriesMapPair.right;
-
-    try {
-      // init QueryDataSource Cache
-      QueryResourceManager.getInstance()
-          .initQueryDataSourceCache(processorToSeriesMap, context, timeFilter);
-    } catch (Exception e) {
-      logger.error("Meet error when init QueryDataSource ", e);
-      throw new QueryProcessException("Meet error when init QueryDataSource.", e);
-    } finally {
-      StorageEngine.getInstance().mergeUnLock(lockList);
-    }
-
-    for (Map.Entry<PartialPath, List<Integer>> entry : pathToAggrIndexesMap.entrySet()) {
-      PartialPath seriesPath = entry.getKey();
-      aggregateOneSeries(
-          seriesPath,
-          entry.getValue(),
-          aggregationPlan.getAllMeasurementsInDevice(seriesPath.getDevice()),
-          timeFilter);
-    }
-    for (Map.Entry<AlignedPath, List<List<Integer>>> entry :
-        alignedPathToAggrIndexesMap.entrySet()) {
-      AlignedPath alignedPath = entry.getKey();
-      aggregateOneAlignedSeries(
-          alignedPath,
-          entry.getValue(),
-          aggregationPlan.getAllMeasurementsInDevice(alignedPath.getDevice()),
-          timeFilter);
-    }
-
-    return constructDataSet(Arrays.asList(aggregateResultList), aggregationPlan);
+    return null;
   }
 
   /**
@@ -621,65 +559,7 @@ public class AggregationExecutor {
   /** execute aggregate function with value filter. */
   public QueryDataSet executeWithValueFilter(AggregationPlan queryPlan)
       throws StorageEngineException, IOException, QueryProcessException {
-    optimizeLastElementFunc(queryPlan);
-
-    TimeGenerator timestampGenerator = getTimeGenerator(context, queryPlan);
-
-    Map<IReaderByTimestamp, List<List<Integer>>> readerToAggrIndexesMap = new HashMap<>();
-
-    // group by path name
-    Map<PartialPath, List<Integer>> pathToAggrIndexesMap =
-        MetaUtils.groupAggregationsBySeries(selectedSeries);
-    Map<AlignedPath, List<List<Integer>>> alignedPathToAggrIndexesMap =
-        MetaUtils.groupAlignedSeriesWithAggregations(pathToAggrIndexesMap);
-
-    List<PartialPath> groupedPathList =
-        new ArrayList<>(pathToAggrIndexesMap.size() + alignedPathToAggrIndexesMap.size());
-    groupedPathList.addAll(pathToAggrIndexesMap.keySet());
-    groupedPathList.addAll(alignedPathToAggrIndexesMap.keySet());
-
-    Pair<List<DataRegion>, Map<DataRegion, List<PartialPath>>> lockListAndProcessorToSeriesMapPair =
-        StorageEngine.getInstance().mergeLock(groupedPathList);
-    List<DataRegion> lockList = lockListAndProcessorToSeriesMapPair.left;
-    Map<DataRegion, List<PartialPath>> processorToSeriesMap =
-        lockListAndProcessorToSeriesMapPair.right;
-
-    try {
-      // init QueryDataSource Cache
-      QueryResourceManager.getInstance()
-          .initQueryDataSourceCache(
-              processorToSeriesMap, context, timestampGenerator.getTimeFilter());
-    } catch (Exception e) {
-      logger.error("Meet error when init QueryDataSource ", e);
-      throw new QueryProcessException("Meet error when init QueryDataSource.", e);
-    } finally {
-      StorageEngine.getInstance().mergeUnLock(lockList);
-    }
-
-    for (PartialPath path : pathToAggrIndexesMap.keySet()) {
-      IReaderByTimestamp seriesReaderByTimestamp =
-          getReaderByTime(path, queryPlan, path.getSeriesType(), context);
-      readerToAggrIndexesMap.put(
-          seriesReaderByTimestamp, Collections.singletonList(pathToAggrIndexesMap.get(path)));
-    }
-    // assign null to be friendly for GC
-    pathToAggrIndexesMap = null;
-    for (AlignedPath vectorPath : alignedPathToAggrIndexesMap.keySet()) {
-      IReaderByTimestamp seriesReaderByTimestamp =
-          getReaderByTime(vectorPath, queryPlan, vectorPath.getSeriesType(), context);
-      readerToAggrIndexesMap.put(
-          seriesReaderByTimestamp, alignedPathToAggrIndexesMap.get(vectorPath));
-    }
-    // assign null to be friendly for GC
-    alignedPathToAggrIndexesMap = null;
-
-    for (int i = 0; i < selectedSeries.size(); i++) {
-      aggregateResultList[i] =
-          AggregateResultFactory.getAggrResultByName(
-              aggregations.get(i), dataTypes.get(i), ascending);
-    }
-    aggregateWithValueFilter(timestampGenerator, readerToAggrIndexesMap);
-    return constructDataSet(Arrays.asList(aggregateResultList), queryPlan);
+    return null;
   }
 
   private void optimizeLastElementFunc(QueryPlan queryPlan) {
