+++ server/src/main/java/org/apache/iotdb/db/query/executor/RawDataQueryExecutor.java
@@ -18,12 +18,8 @@
  */
 package org.apache.iotdb.db.query.executor;
 
-import org.apache.iotdb.commons.path.AlignedPath;
 import org.apache.iotdb.commons.path.MeasurementPath;
 import org.apache.iotdb.commons.path.PartialPath;
-import org.apache.iotdb.db.engine.StorageEngine;
-import org.apache.iotdb.db.engine.querycontext.QueryDataSource;
-import org.apache.iotdb.db.engine.storagegroup.DataRegion;
 import org.apache.iotdb.db.exception.StorageEngineException;
 import org.apache.iotdb.db.exception.query.QueryProcessException;
 import org.apache.iotdb.db.qp.physical.crud.RawDataQueryPlan;
@@ -34,11 +30,9 @@ import org.apache.iotdb.db.query.dataset.RawQueryDataSetWithValueFilter;
 import org.apache.iotdb.db.query.dataset.RawQueryDataSetWithoutValueFilter;
 import org.apache.iotdb.db.query.reader.series.IReaderByTimestamp;
 import org.apache.iotdb.db.query.reader.series.ManagedSeriesReader;
-import org.apache.iotdb.db.query.reader.series.SeriesRawDataBatchReader;
 import org.apache.iotdb.db.query.reader.series.SeriesReaderByTimestamp;
 import org.apache.iotdb.db.query.timegenerator.ServerTimeGenerator;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
-import org.apache.iotdb.tsfile.read.expression.impl.GlobalTimeExpression;
 import org.apache.iotdb.tsfile.read.filter.basic.Filter;
 import org.apache.iotdb.tsfile.read.query.dataset.QueryDataSet;
 import org.apache.iotdb.tsfile.read.query.timegenerator.TimeGenerator;
@@ -49,10 +43,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -105,58 +96,7 @@ public class RawDataQueryExecutor {
 
   protected List<ManagedSeriesReader> initManagedSeriesReader(QueryContext context)
       throws StorageEngineException, QueryProcessException {
-    Filter timeFilter = null;
-    if (queryPlan.getExpression() != null) {
-      timeFilter = ((GlobalTimeExpression) queryPlan.getExpression()).getFilter();
-    }
-
-    List<ManagedSeriesReader> readersOfSelectedSeries = new ArrayList<>();
-    Pair<List<DataRegion>, Map<DataRegion, List<PartialPath>>> lockListAndProcessorToSeriesMapPair =
-        StorageEngine.getInstance().mergeLock(queryPlan.getDeduplicatedPaths());
-    List<DataRegion> lockList = lockListAndProcessorToSeriesMapPair.left;
-    Map<DataRegion, List<PartialPath>> processorToSeriesMap =
-        lockListAndProcessorToSeriesMapPair.right;
-
-    try {
-      // init QueryDataSource cache
-      QueryResourceManager.getInstance()
-          .initQueryDataSourceCache(processorToSeriesMap, context, timeFilter);
-    } catch (Exception e) {
-      logger.error("Meet error when init QueryDataSource ", e);
-      throw new QueryProcessException("Meet error when init QueryDataSource.", e);
-    } finally {
-      StorageEngine.getInstance().mergeUnLock(lockList);
-    }
-
-    try {
-      List<PartialPath> paths = queryPlan.getDeduplicatedPaths();
-      for (PartialPath path : paths) {
-        TSDataType dataType = path.getSeriesType();
-
-        QueryDataSource queryDataSource =
-            QueryResourceManager.getInstance()
-                .getQueryDataSource(path, context, timeFilter, queryPlan.isAscending());
-        timeFilter = queryDataSource.updateFilterUsingTTL(timeFilter);
-
-        ManagedSeriesReader reader =
-            new SeriesRawDataBatchReader(
-                path,
-                queryPlan.getAllMeasurementsInDevice(path.getDevice()),
-                dataType,
-                context,
-                queryDataSource,
-                timeFilter,
-                null,
-                null,
-                queryPlan.isAscending());
-        readersOfSelectedSeries.add(reader);
-      }
-    } catch (Exception e) {
-      logger.error("Meet error when init series reader  ", e);
-      throw new QueryProcessException("Meet error when init series reader .", e);
-    }
-
-    return readersOfSelectedSeries;
+    return null;
   }
 
   /**
@@ -209,81 +149,7 @@ public class RawDataQueryExecutor {
   protected Pair<List<IReaderByTimestamp>, List<List<Integer>>> initSeriesReaderByTimestamp(
       QueryContext context, RawDataQueryPlan queryPlan, List<Boolean> cached, Filter timeFilter)
       throws QueryProcessException, StorageEngineException {
-    List<IReaderByTimestamp> readersOfSelectedSeries = new ArrayList<>();
-
-    List<PartialPath> pathList = new ArrayList<>();
-    List<PartialPath> notCachedPathList = new ArrayList<>();
-
-    // reader index -> deduplicated path index
-    List<List<Integer>> readerToIndexList = new ArrayList<>();
-    // fullPath -> reader index
-    Map<String, Integer> fullPathToReaderIndexMap = new HashMap<>();
-    List<PartialPath> deduplicatedPaths = queryPlan.getDeduplicatedPaths();
-    int index = 0;
-    for (int i = 0; i < cached.size(); i++) {
-      if (cached.get(i)) {
-        pathList.add(deduplicatedPaths.get(i));
-        readerToIndexList.add(Collections.singletonList(i));
-        cached.set(index++, Boolean.TRUE);
-      } else {
-        notCachedPathList.add(deduplicatedPaths.get(i));
-        // For aligned Path, it's deviceID; for nonAligned path, it's full path
-        String fullPath = deduplicatedPaths.get(i).getFullPath();
-        Integer readerIndex = fullPathToReaderIndexMap.get(fullPath);
-
-        // it's another sub sensor in aligned device, we just add it to the previous AlignedPath
-        if (readerIndex != null) {
-          AlignedPath anotherSubSensor = (AlignedPath) deduplicatedPaths.get(i);
-          ((AlignedPath) pathList.get(readerIndex)).mergeAlignedPath(anotherSubSensor);
-          readerToIndexList.get(readerIndex).add(i);
-        } else {
-          pathList.add(deduplicatedPaths.get(i));
-          fullPathToReaderIndexMap.put(fullPath, index);
-          List<Integer> indexList = new ArrayList<>();
-          indexList.add(i);
-          readerToIndexList.add(indexList);
-          cached.set(index++, Boolean.FALSE);
-        }
-      }
-    }
-
-    queryPlan.setDeduplicatedPaths(pathList);
-    int previousSize = cached.size();
-    if (previousSize > pathList.size()) {
-      cached.subList(pathList.size(), previousSize).clear();
-    }
-
-    Pair<List<DataRegion>, Map<DataRegion, List<PartialPath>>> lockListAndProcessorToSeriesMapPair =
-        StorageEngine.getInstance().mergeLock(notCachedPathList);
-    List<DataRegion> lockList = lockListAndProcessorToSeriesMapPair.left;
-    Map<DataRegion, List<PartialPath>> processorToSeriesMap =
-        lockListAndProcessorToSeriesMapPair.right;
-
-    try {
-      // init QueryDataSource Cache
-      QueryResourceManager.getInstance()
-          .initQueryDataSourceCache(processorToSeriesMap, context, timeFilter);
-    } catch (Exception e) {
-      logger.error("Meet error when init QueryDataSource ", e);
-      throw new QueryProcessException("Meet error when init QueryDataSource.", e);
-    } finally {
-      StorageEngine.getInstance().mergeUnLock(lockList);
-    }
-    for (int i = 0; i < queryPlan.getDeduplicatedPaths().size(); i++) {
-      if (cached.get(i)) {
-        readersOfSelectedSeries.add(null);
-        continue;
-      }
-      PartialPath path = queryPlan.getDeduplicatedPaths().get(i);
-      IReaderByTimestamp seriesReaderByTimestamp =
-          getReaderByTimestamp(
-              path,
-              queryPlan.getAllMeasurementsInDevice(path.getDevice()),
-              queryPlan.getDeduplicatedDataTypes().get(i),
-              context);
-      readersOfSelectedSeries.add(seriesReaderByTimestamp);
-    }
-    return new Pair<>(readersOfSelectedSeries, readerToIndexList);
+    return null;
   }
 
   protected IReaderByTimestamp getReaderByTimestamp(
