+++ server/src/test/java/org/apache/iotdb/db/engine/storagegroup/DataRegionTest.java
@@ -30,7 +30,7 @@ import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.constant.TestConstant;
 import org.apache.iotdb.db.engine.MetadataManagerHelper;
-import org.apache.iotdb.db.engine.StorageEngineV2;
+import org.apache.iotdb.db.engine.StorageEngine;
 import org.apache.iotdb.db.engine.compaction.CompactionTaskManager;
 import org.apache.iotdb.db.engine.compaction.inner.InnerSpaceCompactionTask;
 import org.apache.iotdb.db.engine.compaction.log.CompactionLogger;
@@ -90,7 +90,7 @@ public class DataRegionTest {
     MetadataManagerHelper.initMetadata();
     EnvironmentUtils.envSetUp();
     dataRegion = new DummyDataRegion(systemDir, storageGroup);
-    StorageEngineV2.getInstance().setDataRegion(new DataRegionId(0), dataRegion);
+    StorageEngine.getInstance().setDataRegion(new DataRegionId(0), dataRegion);
     CompactionTaskManager.getInstance().start();
   }
 
@@ -98,7 +98,7 @@ public class DataRegionTest {
   public void tearDown() throws Exception {
     if (dataRegion != null) {
       dataRegion.syncDeleteDataFiles();
-      StorageEngineV2.getInstance().deleteDataRegion(new DataRegionId(0));
+      StorageEngine.getInstance().deleteDataRegion(new DataRegionId(0));
     }
     EnvironmentUtils.cleanEnv();
     EnvironmentUtils.cleanDir(TestConstant.OUTPUT_DATA_DIR);
@@ -821,7 +821,7 @@ public class DataRegionTest {
               0);
       CompactionTaskManager.getInstance().addTaskToWaitingQueue(task);
       Thread.sleep(20);
-      List<DataRegion> dataRegions = StorageEngineV2.getInstance().getAllDataRegions();
+      List<DataRegion> dataRegions = StorageEngine.getInstance().getAllDataRegions();
       List<DataRegion> regionsToBeDeleted = new ArrayList<>();
       for (DataRegion region : dataRegions) {
         if (region.getStorageGroupName().equals(storageGroup)) {
@@ -829,7 +829,7 @@ public class DataRegionTest {
         }
       }
       for (DataRegion region : regionsToBeDeleted) {
-        StorageEngineV2.getInstance()
+        StorageEngine.getInstance()
             .deleteDataRegion(new DataRegionId(Integer.parseInt(region.getDataRegionId())));
       }
       Thread.sleep(500);
@@ -870,7 +870,7 @@ public class DataRegionTest {
     long preFLushInterval = config.getSeqMemtableFlushInterval();
     config.setEnableTimedFlushSeqMemtable(true);
     config.setSeqMemtableFlushInterval(5);
-    StorageEngineV2.getInstance().rebootTimedService();
+    StorageEngine.getInstance().rebootTimedService();
 
     Thread.sleep(500);
 
@@ -925,7 +925,7 @@ public class DataRegionTest {
     long preFLushInterval = config.getUnseqMemtableFlushInterval();
     config.setEnableTimedFlushUnseqMemtable(true);
     config.setUnseqMemtableFlushInterval(5);
-    StorageEngineV2.getInstance().rebootTimedService();
+    StorageEngine.getInstance().rebootTimedService();
 
     Thread.sleep(500);
 
@@ -1009,7 +1009,7 @@ public class DataRegionTest {
       }
     }
 
-    List<DataRegion> dataRegions = StorageEngineV2.getInstance().getAllDataRegions();
+    List<DataRegion> dataRegions = StorageEngine.getInstance().getAllDataRegions();
     List<DataRegion> regionsToBeDeleted = new ArrayList<>();
     for (DataRegion region : dataRegions) {
       if (region.getStorageGroupName().equals(storageGroup)) {
@@ -1017,7 +1017,7 @@ public class DataRegionTest {
       }
     }
     for (DataRegion region : regionsToBeDeleted) {
-      StorageEngineV2.getInstance()
+      StorageEngine.getInstance()
           .deleteDataRegion(new DataRegionId(Integer.parseInt(region.getDataRegionId())));
     }
     Thread.sleep(500);
